added more code: Astar -- added declarations with an idea, this is to be edited at a later point.
vehicle and board -- spent more time on, they are being
developed together as vehicle is basically just a subcomponent
of board, and not it's own part.

main -- nothing much here, added Astar as a requirement.

Fails compile due to Astar, and because getPositions still
needs to be solidified on how it should work. flip-flopping
between returning a vector and an array*