status update for july 7:
---- code updates: ----

Astar -- currently commented out, will be
worked upon next

Board -- functionally complete, only the
print and findXCar functions need to be
written at the moment

Vehicle -- completed for now

Main -- currently acting as testing suit
for written code, will be rewritten as
Astar approaches testing state.

---- comments: ----

Today I discovered that
`for(auto Itr : IterableObject){}` was
valid code in the world of C++. To find
out that C++ has a for loop that is similar
to that one would find in python is amazing,
and I will be using it from now on in
situations where I need to be iterating
through easily iterable objects.