status update for 8 July:
---- code updates: ----

Astar -- wrote generateNewBoards, which
will call upon all the slide functions.
These have yet to be written.

Board -- fixed a minor error in addVehicle.
also added the capacity to return a list of
vehicle names, this is necessary for A*
functionality. Removed a const, as to allow
for the copy function to work out of the box.

Vehicle -- Added a getName function, and
removed the const from the header to allow
for default copy constructor to work.

Main -- removed previous testing code,
and added functionality to read in a
new gameboard.

---- comments: ----

Will need to look in abort() more
as I have not yet made a proper error
handling mechanism, and this is small
and simple enough that if there is an
error, it could be handled simply by
a SIGABORT. possibly.